+++ DronPlan.Core/Domain/Model/MDrone.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int MaxWeight { get; set; }
+        public double MaxWeight { get; set; }
 
         public List<List<MPackage>> Trips { get; set; }
     }
